Another batch of 128-bit decimal optimizations

Benchmarks (BenchmarkDecimalOperators class):
Benchmark   | throughput before | throughput after
-|-|-
s2 + l2 | 557 | 705
l1 + l2 | 571 | 601
l1 + l2 + l3 + l4 | 240 | 262
s2 + l3 + l1 + s4 | 251 | 307
s2 + lz3 + lz1 + s4 | 230 | 285
s1 / s3 | 265 | 334
s2 / l1 | 391 | 497
l1 / s2 | 279 | 354
s3 / l1 | 224 | 294
l2 / l3 | 186 | 222
l2 / l4 / l4 / l4 | 113 | 128
l2 / s4 / s4 / s4 | 110 | 140
s2 % l2 | 443 | 578
l3 % s3 | 272 | 358
s4 % l3 | 320 | 379
l2 % l3 | 260 | 299
l2 % l3 % l4 % l1 | 91 | 113
s3 * s4 | 586 | 1153
l2 * s2 | 716 | 855
l2 * s2 * s5 * s6 | 329 | 448
s1 * l2 | 713 | 865