@skrzypo987 2020-09-18
I highly appreciate two thumbs up and 3 hearts. However code-review would also be great.