@skrzypo987 2020-10-02
Since this is still pending for review I added 4 new commits.
I will post new benchmark results in a few hours.