@skrzypo987 2020-10-12
ping @sopel39 
26. day seems like a perfect time to post a remainder about the review.