@skrzypo987 2020-10-20
@sopel39 
Remarks addressed.
Added one commit with static imports (the first one)